[EST-68] Refactor: 함수이름 수정
